requirements.txt

# Brain-Heart Deep Research System Requirements
# Comprehensive dependencies for all search and social media providers

# Core dependencies (Required)
streamlit>=1.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.28.0
fastapi

# LLM providers (choose based on your needs)
openai>=1.0.0
anthropic>=0.7.0
groq>=0.4.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# RAG DEPENDENCIES 
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai
PyJWT>=2.0.0

# PDF and DOCX support for LangChain
pypdf>=3.0.0
unstructured
unstructured[ppt]

#google Oauth
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

#encryption
cryptography
# Google APIs (for Google CSE, YouTube)
google-api-python-client>=2.70.0
google-auth>=2.15.0

# Social media APIs
praw>=7.6.0  # Reddit API

# Web scraping (for Instagram/Facebook)
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Optional: Enhanced functionality
nltk>=3.8.0
textblob>=0.17.0
matplotlib>=3.6.0

# Development (optional)
pytest>=7.0.0
black>=22.0.0
python-json-logger
python-multipart

mem0ai
langchain-neo4j
rank-bm25

unstructured[md]
fastapi
fastapi-limiter
redis
pymongo
pypdf2
trafilatura
piexif

# Subprocess MCP server runtime used by redis-mcp-server (provides `uvx` CLI)
# Install `uv` so the `uvx` command is available for Redis MCP integration.
uv>=0.7.0
